feat: added pipeline info test and seperated func
