+++ packages/developer-portal/src/components/pages/v8/__tests__/pipeline-info.test.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+import { shallow } from 'enzyme'
+import { PipelineInfo } from '../pipeline/pipeline-info'
+
+describe('Pipeline info', () => {
+  it('Should match snapshot', () => {
+    expect(
+      shallow(
+        <PipelineInfo
+          pipeline={{
+            id: 'pipelineId',
+            buildStatus: 'IN_PROGRESS',
+          }}
+        />,
+      ),
+    ).toMatchSnapshot()
+  })
+})
