+++ packages/developer-portal/src/components/pages/v8/pipeline/deployment-info.tsx
@@ -1,18 +1,6 @@
 import { cx } from '@linaria/core'
 import { useReapitConnect } from '@reapit/connect-session'
-import {
-  BodyText,
-  Button,
-  ButtonGroup,
-  ColSplit,
-  elMb6,
-  InputWrap,
-  Intent,
-  StatusIndicator,
-  Subtitle,
-  Title,
-  Grid,
-} from '@reapit/elements'
+import { Button, ButtonGroup, elMb6, Title } from '@reapit/elements'
 import { PipelineModelInterface, PipelineRunnerModelInterface } from '@reapit/foundations-ts-definitions'
 import { GetActionNames, getActions, UpdateActionNames, updateActions } from '@reapit/utils-common'
 import { useReapitGet, useReapitUpdate } from '@reapit/utils-react'
@@ -21,67 +9,12 @@ import { reapitConnectBrowserSession } from '../../../../core/connect-session'
 import { openNewPage } from '@/utils/navigation'
 import { UpdateReturnTypeEnum } from '@reapit/utils-react/src/use-reapit-data/use-reapit-update'
 import { PipelineDeploymentTable } from './pipeline-runner-table'
+import { PipelineInfo } from './pipeline-info'
 
-const buildStatusToIntent = (status: string): Intent => {
-  switch (status) {
-    case 'CREATING_ARCHITECTURE':
-      return 'primary'
-    case 'SUCCESS':
-      return 'primary'
-    case 'RUNNING':
-      return 'secondary'
-    case 'PENDING':
-    case 'QUEUED':
-    default:
-      return 'neutral'
-  }
-}
-
-const buildStatusToReadable = (status: string): string =>
-  status
-    .split('_')
-    .map((str) => str.toLowerCase())
-    .join(' ')
-
-const PipelineInfo: React.FC<{ pipeline: PipelineModelInterface }> = ({ pipeline }) => {
-  return (
-    <Grid>
-      <ColSplit>
-        <InputWrap>
-          <Subtitle>Status</Subtitle>
-          <BodyText>
-            <StatusIndicator intent={buildStatusToIntent(pipeline.buildStatus as string)} />{' '}
-            {buildStatusToReadable(pipeline.buildStatus as string)}
-          </BodyText>
-        </InputWrap>
-        <InputWrap>
-          <Subtitle>Repository</Subtitle>
-          <BodyText>{pipeline.repository}</BodyText>
-        </InputWrap>
-        <InputWrap>
-          <Subtitle>Package Manager</Subtitle>
-          <BodyText>{pipeline.packageManager}</BodyText>
-        </InputWrap>
-      </ColSplit>
-      <ColSplit>
-        <InputWrap>
-          <Subtitle>Build Command</Subtitle>
-          <BodyText>{pipeline.buildCommand}</BodyText>
-        </InputWrap>
-        <InputWrap>
-          <Subtitle>Tests</Subtitle>
-          <BodyText>{pipeline.testCommand}</BodyText>
-        </InputWrap>
-        <InputWrap>
-          <Subtitle>Location</Subtitle>
-          <BodyText>{pipeline.subDomain ? `https://${pipeline.subDomain}.dev.paas.reapit.cloud` : ''}</BodyText>
-        </InputWrap>
-      </ColSplit>
-    </Grid>
-  )
-}
-
-export const PipelineDeploymentInfo = ({ pipeline, channel }: { pipeline: PipelineModelInterface; channel: any }) => {
+export const PipelineDeploymentInfo: React.FC<{ pipeline: PipelineModelInterface; channel: any }> = ({
+  pipeline,
+  channel,
+}) => {
   const { connectSession } = useReapitConnect(reapitConnectBrowserSession)
   const [pipelineDeployments, loading] = useReapitGet<{ items: PipelineRunnerModelInterface[] }>({
     reapitConnectBrowserSession,
